Use context conversational retrieval QA prompt only in docs
