+++ langchain/src/chains/tests/conversational_retrieval_chain.int.test.ts
@@ -330,99 +330,3 @@ test("Test ConversationalRetrievalQAChain from LLM with a chat model and memory"
 
   console.log({ res2 });
 });
-
-test("Test ConversationalRetrievalQAChain from LLM with a chat model, memory, and keeping the chat history context", async () => {
-  const model = new ChatOpenAI({
-    modelName: "gpt-3.5-turbo",
-    temperature: 0,
-  });
-
-  const vectorStore = await HNSWLib.fromTexts(
-    [
-      "Mitochondria are the powerhouse of the cell",
-      "Foo is red",
-      "Bar is red",
-      "Buildings are made out of brick",
-      "Mitochondria are made of lipids",
-    ],
-    [{ id: 2 }, { id: 1 }, { id: 3 }, { id: 4 }, { id: 5 }],
-    new OpenAIEmbeddings()
-  );
-
-  const chain = ConversationalRetrievalQAChain.fromLLM(
-    model,
-    vectorStore.asRetriever(),
-    {
-      memory: new BufferMemory({
-        memoryKey: "chat_history",
-        returnMessages: true,
-      }),
-      questionGeneratorChainOptions: {
-        extractConversationContext: true,
-      },
-    }
-  );
-
-  await chain.call({
-    question:
-      "I have a friend called Bob. He's 28 years old. He'd like to know what is the powerhouse of the cell?",
-  });
-
-  const res2 = await chain.call({
-    question: "How old is Bob?",
-  });
-
-  expect(res2).toEqual(
-    expect.objectContaining({
-      text: expect.stringContaining("28"),
-    })
-  );
-});
-
-test("Test ConversationalRetrievalQAChain from LLM with a chat model, memory, and without keeping the chat history context", async () => {
-  const model = new ChatOpenAI({
-    modelName: "gpt-3.5-turbo",
-    temperature: 0,
-  });
-
-  const vectorStore = await HNSWLib.fromTexts(
-    [
-      "Mitochondria are the powerhouse of the cell",
-      "Foo is red",
-      "Bar is red",
-      "Buildings are made out of brick",
-      "Mitochondria are made of lipids",
-    ],
-    [{ id: 2 }, { id: 1 }, { id: 3 }, { id: 4 }, { id: 5 }],
-    new OpenAIEmbeddings()
-  );
-
-  const chain = ConversationalRetrievalQAChain.fromLLM(
-    model,
-    vectorStore.asRetriever(),
-    {
-      memory: new BufferMemory({
-        memoryKey: "chat_history",
-        returnMessages: true,
-      }),
-      questionGeneratorChainOptions: {
-        extractConversationContext: false,
-      },
-    }
-  );
-
-  await chain.call({
-    question:
-      "I have a friend called Bob. He's 28 years old. He'd like to know what is the powerhouse of the cell?",
-  });
-
-  const res2 = await chain.call({
-    question: "How old is Bob?",
-  });
-
-  expect(res2).toEqual(
-    expect.objectContaining({
-      text: expect.not.stringMatching("28"),
-    })
-  );
-});
